feat: add book custom exceptions

+++ backend/src/main/java/br/com/bookly/services/impl/BookServiceImpl.java
@@ -1,8 +1,11 @@
 package br.com.bookly.services.impl;
 
 import br.com.bookly.entities.Book;
+import br.com.bookly.exceptions.InexistentAuthorException;
+import br.com.bookly.exceptions.InexistentBookException;
+import br.com.bookly.exceptions.InvalidBookDataException;
 import br.com.bookly.repositories.BookRepository;
-import br.com.bookly.services.bookService;
+import br.com.bookly.services.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -12,7 +15,7 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 @Service
-public class BookServiceImpl implements bookService {
+public class BookServiceImpl implements BookService {
 
     @Autowired
     BookRepository bookRepository;
@@ -26,23 +29,23 @@ public class BookServiceImpl implements bookService {
     public Book addBook(Book book) {
 
         if(book.getTitle() == null || book.getTitle().isBlank()) {
-            return null;
+            throw new InvalidBookDataException("Error: Empty Book Title");
         }
 
         if(book.getSynopsis() == null || book.getSynopsis().isBlank()) {
-            return null;
+            throw new InvalidBookDataException("Error: Empty Book Synopsis");
         }
 
         if(book.getGenre() == null || book.getGenre().isBlank()) {
-            return null;
+            throw new InvalidBookDataException("Error: Empty Book Genre");
         }
 
         if(book.getAvailableQuantity() < 0) {
-            return null;
+            throw new InvalidBookDataException("Error: Invalid Book Available Quantity (Negative Value)");
         }
 
         if(book.getPrice().compareTo(BigDecimal.ZERO) <= 0) {
-            return null;
+            throw new InvalidBookDataException("Error: Invalid Book Price (Negative Value)");
         }
 
         return bookRepository.save(book);
@@ -53,13 +56,13 @@ public class BookServiceImpl implements bookService {
     public Book updateBook(UUID id, Book book) {
 
         if(book == null || id == null){
-            return null;
+            throw new InvalidBookDataException("Error: Invalid Book or Book ID");
         }
 
         Book bookById = bookRepository.findById(id).orElse(null);
 
         if(bookById == null) {
-            return null;
+            throw new InexistentBookException();
         }
 
         bookById.setTitle(book.getTitle());
@@ -73,6 +76,12 @@ public class BookServiceImpl implements bookService {
 
     @Override
     public boolean deleteBook(UUID id) {
+        Book bookById = bookRepository.findById(id).orElse(null);
+
+        if(bookById == null) {
+            throw new InexistentBookException();
+        }
+
         bookRepository.deleteById(id);
 
         if(!bookRepository.existsById(id)){
@@ -85,14 +94,18 @@ public class BookServiceImpl implements bookService {
     @Override
     public Book increaseAvaliableBooks(UUID id, int amount) {
 
-        if(id == null || amount <= 0){
-            return null;
+        if(id == null){
+            throw new InvalidBookDataException("Error: Null Book ID");
+        }
+
+        if(amount <= 0){
+            throw new InvalidBookDataException("Error: Invalid Book Amount to Increase (Negative Value)");
         }
 
         Book bookById = bookRepository.findById(id).orElse(null);
 
         if(bookById == null) {
-            return null;
+            throw new InexistentBookException();
         }
 
         bookById.setAvailableQuantity(bookById.getAvailableQuantity() + amount);
@@ -104,17 +117,17 @@ public class BookServiceImpl implements bookService {
     @Override
     public Book decreaseAvaliableBooks(UUID id, int amount) {
         if(id == null){
-            return null;
+            throw new InvalidBookDataException("Error: Null Book ID");
         }
 
         Book bookById = bookRepository.findById(id).orElse(null);
 
         if(bookById == null) {
-            return null;
+            throw new InexistentBookException();
         }
 
         if(bookById.getAvailableQuantity() < amount){
-            return null;
+            throw new InvalidBookDataException("Error: Amount to Decrease Greater Than Available Quantity)");
         }
 
         bookById.setAvailableQuantity(bookById.getAvailableQuantity() - amount);
